Prepare to start interpreter runtime builder.
