+++ src/main/java/org/joshsim/lang/interpret/InterpreterContext.java
@@ -1,0 +1,15 @@
+/**
+ * Structures describing a scope for an InterpreterAction.
+ *
+ * @license BSD-3-Clause
+ */
+
+package org.joshsim.lang.interpret;
+
+
+/**
+ * Structure which describes the context in which an interpreter action is taken.
+ */
+public interface InterpreterContext {
+
+}
