+++ src/main/java/org/joshsim/lang/interpret/InterpreterRuntimeBuilder.java
@@ -1,0 +1,15 @@
+/**
+ * Structures describing an interpreter runtime while be built.
+ *
+ * @license BSD-3-Clause
+ */
+
+package org.joshsim.lang.interpret;
+
+
+/**
+ * Structure which helps build an interpreter runtime.
+ */
+public interface InterpreterRuntimeBuilder {
+
+}
